+++ resources/assets/css/style.scss
@@ -44,6 +44,17 @@ body {
   }
 }
 
+.menu-list {
+  & li {
+    font-size: 14px;
+    margin-bottom: 3px;
+    cursor: pointer;
+    &:hover {
+      opacity: 0.6;
+    }
+  }
+}
+
 /* Bulma 上書き対応 */
 .card ul.article__category {
   margin: 0;
@@ -65,16 +76,5 @@ body {
   }
   &__content {
     max-width: 1000px;
-  }
-}
-
-.menu-list {
-  & li {
-    font-size: 14px;
-    margin-bottom: 3px;
-    cursor: pointer;
-    &:hover {
-      opacity: 0.6;
-    }
   }
 }
